requirements.txt
openai-whisper
numpy
pydantic
piper-tts
sounddevice
scipy
PyQt5
flask
